requirements.txt
strands-agents
strands-agents-tools
prefect==2.14.0
python-dotenv==1.0.0
requests